Spremenil dokoncano v integer in dodal progress bar

+++ models.py
@@ -6,7 +6,7 @@ class Task (ndb.Model):
     dodan = ndb.TimeProperty(auto_now_add=True)
     naslov = ndb.StringProperty()
     besedilo = ndb.TextProperty()
-    dokoncano = ndb.FloatProperty(default=0.0)
+    dokoncano = ndb.IntegerProperty(default=0)
     aktiven = ndb.BooleanProperty(default=True)
     pomembnost = ndb.IntegerProperty()
